Added proper documentation for Search use case, refactored response formatter to output boundary
