--- src/main/java/search_use_case/SearchFail.java
@@ -1,7 +1,0 @@
-package search_use_case;
-
-public class SearchFail extends RuntimeException{
-    public SearchFail(String error){
-        super(error);
-    }
-}
