+++ src/main/java/search_use_case/SearchPresenter.java
@@ -1,8 +1,20 @@
 package search_use_case;
+/**
+ * The search presenter.
+ * <p>
+ * Interface adapters.
+ * @author Winston Chieng
+ */
+public class SearchPresenter implements SearchOutputBoundary {
 
-public interface SearchPresenter {
+    @Override
+    public SearchResponseModel prepareSuccessView(SearchResponseModel results) {
+        return results;
+    }
 
-    SearchResponseModel prepareSuccessView(SearchResponseModel results);
-
-    SearchResponseModel prepareFailView(String error);
+    @Override
+    public SearchResponseModel prepareFailView(String error) {
+        throw new SearchFailScreen(error);
+    }
 }
+
