+++ src/main/java/search_use_case/SearchRequestModel.java
@@ -3,39 +3,54 @@ import entityRequestModels.CommonUserDsRequestModel;
 
 import java.util.ArrayList;
 
+/**
+ * A data structure containing the search input
+ * from the user, the selected tags, and the
+ * current user searching.
+ * <p>
+ * Application Business Rules.
+ * @author Winston Chieng
+ */
 public class SearchRequestModel {
 
-    private String search_input;
-    private ArrayList<String> tags;
-    private CommonUserDsRequestModel user;
+    private final String search_input;
+    private final ArrayList<String> tags;
+    private final CommonUserDsRequestModel user;
 
+    /**
+     *  creates a SearchRequestModel
+     * @param search_input the user input
+     * @param tags  the selected tags
+     * @param user  the current user
+     */
     public SearchRequestModel(String search_input, ArrayList<String> tags, CommonUserDsRequestModel user){
         this.search_input = search_input;
         this.tags = tags;
         this.user = user;
     }
 
+    /**
+     *
+     * @return the search input
+     */
     String getSearch_input(){
         return search_input;
     }
-    void setSearch_input(String search_input){
-        this.search_input = search_input;
-    }
 
+    /**
+     *
+     * @return the ArrayList of tags selected by the user
+     */
     ArrayList<String> getTags(){
         return tags;
     }
 
-    void setTags(ArrayList<String> tags){
-        this.tags = tags;
-    }
-
+    /**
+     *
+     * @return the current user
+     */
     CommonUserDsRequestModel getUser(){
         return user;
-    }
-
-    void setUser(CommonUserDsRequestModel user){
-        this.user = user;
     }
 
 }
